requirements.txt
argparse
GitPython